% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMB.R
\name{UMB}
\alias{UMB}
\title{Unit Maxwell-Boltzmann family}
\usage{
UMB(mu.link = "log")
}
\arguments{
\item{mu.link}{defines the mu.link, with "log" link as the default for the mu.}
}
\value{
Returns a gamlss.family object which can be used to fit a UMB
distribution in the \code{gamlss()} function.
}
\description{
The function \code{UMB()} defines the Unit Maxwell-Boltzmann distribution, a one parameter
distribution, for a \code{gamlss.family} object to be used in GAMLSS fitting
using the function \code{gamlss()}.
}
\details{
The Unit Maxwell-Boltzmann distribution with parameter \eqn{\mu}
has a support in \eqn{(0, 1)} and density given by

\eqn{f(x| \mu) = \frac{\sqrt(2/\pi) \log^2(1/x) \exp(-\frac{\log^2(1/x)}{2\mu^2})}{\mu^3 x} }

for \eqn{0 < x < 1} and \eqn{\mu > 0}.
}
\examples{
# Example 1
# Generating some random values with
# known mu
y <- rUMB(n=300, mu=0.5)

# Fitting the model
library(gamlss)
mod1 <- gamlss(y~1, family=UMB,
               control=gamlss.control(n.cyc=500, trace=FALSE))

# Extracting the fitted values for mu
# using the inverse link function
exp(coef(mod1, what="mu"))

# Example 2
# Generating random values under some model

# A function to simulate a data set with Y ~ UMB
gendat <- function(n) {
  x1 <- runif(n)
  mu <- exp(-0.5 + 1 * x1)
  y <- rUMB(n=n, mu=mu)
  data.frame(y=y, x1=x1)
}

datos <- gendat(n=300)

mod2 <- gamlss(y~x1,
               family=UMB, data=datos,
               control=gamlss.control(n.cyc=500, trace=TRUE))
summary(mod2)

}
\references{
Biçer, C., Bakouch, H. S., Biçer, H. D., Alomair, G.,
Hussain, T., y Almohisen, A. (2024).
Unit Maxwell-Boltzmann Distribution and Its Application to
Concentrations Pollutant Data.
Axioms, 13(4), 226.
}
\seealso{
\link{dUMB}
}
