% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UPHN.R
\name{UPHN}
\alias{UPHN}
\title{The Unit-Power Half-Normal family}
\usage{
UPHN(mu.link = "log", sigma.link = "log")
}
\arguments{
\item{mu.link}{defines the mu.link, with "log" link as the default for the mu parameter.}

\item{sigma.link}{defines the sigma.link, with "log" link as the default for the sigma.}
}
\value{
Returns a \code{gamlss.family} object which can be used
to fit a COMPO distribution
in the \code{gamlss()} function.
}
\description{
The function \code{UPHN()} defines the Unit-Power Half-Normal
distribution, a two parameter
distribution, for a \code{gamlss.family} object to be used in GAMLSS fitting
using the function \code{gamlss()}.
}
\details{
The UPHN distribution with parameters
\eqn{\mu} and \eqn{\sigma}
has a support in \eqn{(0, 1)} and density given by

\eqn{f(x| \mu, \sigma) = \frac{2\mu}{\sigma x^2} \phi(\frac{1-x}{\sigma x}) (2 \Phi(\frac{1-x}{\sigma x})-1)^{\mu-1}}

for \eqn{0 < x < 1}, \eqn{\mu > 0} and \eqn{\sigma > 0}.
}
\examples{
# Example 1
# Generating random values with
# known mu and sigma
require(gamlss)
mu <- 1.5
sigma <- 4.0

y <- rUPHN(1000, mu, sigma)

mod1 <- gamlss(y~1, sigma.fo=~1,  family=UPHN,
               control=gamlss.control(n.cyc=5000, trace=TRUE))

exp(coef(mod1, what="mu"))
exp(coef(mod1, what="sigma"))

# Example 2
# Generating random values under some model

# A function to simulate a data set with Y ~ UPHN
gendat <- function(n) {
  x1 <- runif(n)
  x2 <- runif(n)
  mu <- exp(0.75 - 0.69 * x1)   # Approx 1.5
  sigma <- exp(0.5 - 0.64 * x2) # Approx 1.20
  y <- rUPHN(n, mu, sigma)
  data.frame(y=y, x1=x1, x2=x2)
}

dat <- gendat(n=2000)

mod2 <- gamlss(y~x1, sigma.fo=~x2, family=UPHN, data=dat,
               control=gamlss.control(n.cyc=5000, trace=TRUE))

summary(mod2)
}
\references{
Santoro, K. I., Gómez, Y. M., Soto, D., & Barranco-Chamorro, I. (2024).
Unit-Power Half-Normal Distribution Including Quantile Regression
with Applications to Medical Data. Axioms, 13(9), 599.
}
\seealso{
\link{dUPHN}.
}
\author{
Juan Diego Suarez Hernandez, \email{jsuarezhe@unal.edu.co}
}
